Roulette done, moving to engine
Randomized Roulette wheel probability, updated the mutation so
that mutation occurs on every crossover, the probability changes
the number of bits flipped.  Need to remove replacement on bit flippin

Next step is updating the main script to run X # of generations on every
one of the models. Done by pulling all GA code out into a function
and re-parameterizing the model.